+++ lib/features/common/views/splash_screen.dart
@@ -1,6 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:stuverse/app/app.dart';
 
+import '../widgets/bg_greadient.dart';
+
 class SplashScreen extends StatelessWidget {
   const SplashScreen({super.key});
 
@@ -23,27 +25,6 @@ class SplashScreen extends StatelessWidget {
           ),
         ],
       )),
-    );
-  }
-}
-
-class BgGradient extends StatelessWidget {
-  const BgGradient({
-    super.key,
-    required this.child,
-  });
-
-  final Widget child;
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      width: MediaQuery.of(context).size.width, // 100%
-      height: MediaQuery.of(context).size.height,
-      decoration: BoxDecoration(
-        gradient: AppTheme.backgroundGradient,
-      ),
-      child: child,
     );
   }
 }
